backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pydantic==2.11.7
pydantic-settings==2.10.1
numpy==2.3.2
pandas==2.3.1
scikit-learn==1.7.1
joblib==1.5.1
httpx==0.28.1
pypdf==5.9.0
SQLAlchemy==2.0.39
asyncpg==0.30.0
aiosqlite==0.21.0
Authlib==1.6.0
PyJWT==2.10.1
itsdangerous==2.2.0
